# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_controller.createServerReq:
    properties:
      auth_type:
        type: string
      ip:
        type: string
      name:
        description: |-
          createServerReq 为创建服务器的请求体。
          说明：password/private_key 为明文输入，仅用于传输；后端入库会进行加密存储。
        type: string
      password:
        type: string
      port:
        type: integer
      private_key:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
...
  k8s-platform-backend_internal_service.FileContent:
    properties:
      content:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
  k8s-platform-backend_internal_service.FileEntry:
    properties:
      group:
        type: string
      is_dir:
        type: boolean
      mod_time:
        description: RFC3339
        type: string
      mode:
        description: e.g. "-rwxr-xr-x"
        type: string
      name:
        type: string
      owner:
        type: string
      path:
        type: string
      size:
        type: integer
      user:
        type: string
    type: object
...
  k8s-platform-backend_pkg_resp.Result:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 星枢K8S管理平台后端接口文档（统一响应结构：code/message/data）
  title: 星枢K8S管理平台 API
  version: "1.0"
paths:
...
  /servers:
    get:
      consumes:
      - application/json
      description: 分页列出服务器，支持 keyword/tag/status 过滤与排序
      parameters:
      - description: 页码
        example: 1
        in: query
        name: page
        type: integer
      - description: 每页条数
        example: 10
        in: query
        name: page_size
        type: integer
      - description: 关键字（name/ip 模糊搜索）
        in: query
        name: keyword
        type: string
      - description: 标签过滤
        in: query
        name: tag
        type: string
      - description: 状态过滤
        example: active
        in: query
        name: status
        type: string
      - description: 排序字段
        example: created_at
        in: query
        name: sort_by
        type: string
      - description: 排序方向
        example: desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询失败
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 服务器列表
      tags:
      - 服务器管理接口
    post:
      consumes:
      - application/json
      description: 创建服务器资产（敏感认证信息入库前会加密）
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controller.createServerReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建失败
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 创建服务器
      tags:
      - 服务器管理接口
  /servers/{id}:
    delete:
      consumes:
      - application/json
      description: 根据服务器 ID 软删除
      parameters:
      - description: 服务器ID
        example: 1001
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除失败
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 删除服务器
      tags:
      - 服务器管理接口
    get:
      consumes:
      - application/json
      description: 根据服务器 ID 查询详情
      parameters:
      - description: 服务器ID
        example: 1001
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询失败
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 服务器详情
      tags:
      - 服务器管理接口
    patch:
      consumes:
      - application/json
      description: 根据服务器 ID 进行部分字段更新
      parameters:
      - description: 服务器ID
        example: 1001
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 更新失败
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 更新服务器
      tags:
      - 服务器管理接口
...
  /servers/{id}/files:
    delete:
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文件路径
        in: query
        name: path
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 删除文件
      tags:
      - 服务器文件管理
    get:
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      - description: 远程路径（默认 home 目录）
        in: query
        name: path
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/k8s-platform-backend_internal_service.FileEntry'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 列出远程目录
      tags:
      - 服务器文件管理
  /servers/{id}/files/download:
    get:
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文件路径
        in: query
        name: path
        required: true
        type: string
      responses: {}
      security:
      - BearerAuth: []
      summary: 下载文件
      tags:
      - 服务器文件管理
  /servers/{id}/files/mkdir:
    post:
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 创建目录
      tags:
      - 服务器文件管理
  /servers/{id}/files/read:
    get:
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文件路径
        in: query
        name: path
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/k8s-platform-backend_internal_service.FileContent'
              type: object
      security:
      - BearerAuth: []
      summary: 读取远程文件
      tags:
      - 服务器文件管理
  /servers/{id}/files/rename:
    post:
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 重命名文件
      tags:
      - 服务器文件管理
  /servers/{id}/files/stat:
    get:
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文件路径
        in: query
        name: path
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/k8s-platform-backend_internal_service.FileEntry'
              type: object
      security:
      - BearerAuth: []
      summary: 文件信息
      tags:
      - 服务器文件管理
  /servers/{id}/files/upload:
    post:
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      - description: 远程目标路径
        in: formData
        name: path
        required: true
        type: string
      - description: 上传的文件
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - 服务器文件管理
  /servers/{id}/files/write:
    put:
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 写入远程文件
      tags:
      - 服务器文件管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"